Backend/requirements.txt
flask
embedchain
python-dotenv